requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
opencv-python>=4.8.0
mediapipe>=0.10.0
facenet-pytorch>=2.5.3
albumentations>=1.3.1
numpy>=1.24.0
Pillow>=9.5.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0
wandb>=0.15.0
PyYAML>=6.0
tqdm>=4.65.0
timm>=0.9.0
